requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
gehomesdk>=0.4.8
firebase-admin>=6.4.0
apscheduler>=3.10.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
cryptography>=42.0.0
python-jose>=3.3.0
python-roborock>=0.12.0
